split supervisor into different files
